+++ src/main/go/inttest/util/assert_db.go
@@ -1,7 +1,6 @@
 package util
 
 import (
-	"database/sql"
 	inttestinfra "github.com/benizzio/open-asset-allocator/inttest/infra"
 	dbx "github.com/go-ozzo/ozzo-dbx"
 	"github.com/stretchr/testify/assert"
@@ -19,16 +18,4 @@ func AssertDBWithQuery(t *testing.T, sql string, expected dbx.NullStringMap) {
 	}
 
 	assert.Equal(t, expected, actual, "Query result does not match expected value")
-}
-
-func ToNullString(str string) sql.NullString {
-	return ToNullStringFromPointer(&str)
-}
-
-func ToNullStringFromPointer(str *string) sql.NullString {
-	if str == nil {
-		return sql.NullString{Valid: false}
-	} else {
-		return sql.NullString{String: *str, Valid: true}
-	}
 }
